gcp/healthcare-data-compliance-workflows-healthcare-api-tasks/code/terraform/function_templates/requirements.txt
google-cloud-healthcare==2.15.0
google-cloud-tasks==2.16.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.1
google-cloud-logging==3.8.0
google-cloud-bigquery==3.13.0
functions-framework==3.4.0